Disable evdev calibration on relative input
